Fixed 64-bit warnings.

+++ ViewDeck/IIViewDeckController.h
@@ -74,7 +74,7 @@ enum {
     IIViewDeckTopSide = 3,
     IIViewDeckBottomSide = 4,
 };
-typedef UInt32 IIViewDeckSide;
+typedef NSInteger IIViewDeckSide;
 
 enum {
     IIViewDeckHorizontalOrientation = 1,
